+++ src/edit/dateFields.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useMemo } from 'react';
 import PropTypes from 'prop-types';
 import { Field } from 'react-final-form';
 import { useIntl } from 'react-intl';
@@ -12,12 +12,24 @@ import {
 
 import { DATE_LENGTH } from '../validation';
 
-const DateFields = ({ instanceDateTypeOptions, disabled }) => {
+const DateFields = ({ instanceDateTypeOptions, disabled, initialDateTypeId }) => {
   const { formatMessage } = useIntl();
 
   const typeLabel = formatMessage({ id: 'ui-inventory.dateType' });
   const date1Label = formatMessage({ id: 'ui-inventory.date1' });
   const date2Label = formatMessage({ id: 'ui-inventory.date2' });
+
+  // if there is no initial date type value - means it's not been saved so a user can unselect it
+  // once a value is saved - user cannot unselect it, but they can still select "No attempt to code"
+  const canUnselectDateType = !initialDateTypeId;
+  const dataOptions = useMemo(() => {
+    return canUnselectDateType
+      ? [{
+        label: formatMessage({ id: 'ui-inventory.selectInstanceDateType' }),
+        value: '',
+      }, ...instanceDateTypeOptions]
+      : instanceDateTypeOptions;
+  }, [initialDateTypeId]);
 
   return (
     <Row>
@@ -26,7 +38,8 @@ const DateFields = ({ instanceDateTypeOptions, disabled }) => {
           label={typeLabel}
           name="dates.dateTypeId"
           component={Select}
-          dataOptions={instanceDateTypeOptions}
+          dataOptions={dataOptions}
+          placeholder={canUnselectDateType ? null : formatMessage({ id: 'ui-inventory.selectInstanceDateType' })}
           disabled={disabled}
         />
       </Col>
@@ -61,6 +74,7 @@ DateFields.propTypes = {
     selected: PropTypes.bool,
   })).isRequired,
   disabled: PropTypes.bool,
+  initialDateTypeId: PropTypes.string,
 };
 
 export default DateFields;